* file: read() , write() return size_t instead of signed int
git-svn-id: https://svn.gwserver.de:3691/hddwiper/trunk@30 c257ac6a-9548-4c2e-97fb-1c9d84f8ff2e

+++ file/InputFile.hpp
@@ -12,7 +12,7 @@ class InputFile: public AbstractFile
 public:
 	InputFile(const std::string &filename);
 
-	int read(Data &data);
+	size_t read(Data &data);
 	
 	unsigned long long int getBytesRead() const;
